+++ src/ui/control/geolocate_control.test.ts
@@ -583,4 +583,13 @@ describe('GeolocateControl with no options', () => {
         await zoomendPromise;
         expect(geolocate._circleElement.style.width).toBeTruthy();
     });
+
+    test('don\'t show button if showButton = false', async () => {
+        const geolocate = new GeolocateControl({
+            showButton: false
+        });
+        map.addControl(geolocate);
+        await sleep(0);
+        expect(geolocate._geolocateButton.classList.contains('maplibregl-ctrl-geolocate-hidden')).toBeTruthy();
+    });
 });